chore(progress-bar): clean up storybook
